Remove legacy settings panel.
