+++ OpenRA.Mods.Cnc/OpenRA.Mods.Cnc.csproj
@@ -94,7 +94,6 @@
     <Compile Include="Widgets\Logic\CncIngameMenuLogic.cs" />
     <Compile Include="Widgets\Logic\CncInstallFromCDLogic.cs" />
     <Compile Include="Widgets\Logic\CncMenuLogic.cs" />
-    <Compile Include="Widgets\Logic\CncSettingsLogic.cs" />
     <Compile Include="Widgets\Logic\ProductionTooltipLogic.cs" />
     <Compile Include="Widgets\Logic\SupportPowerTooltipLogic.cs" />
     <Compile Include="Widgets\ProductionPaletteWidget.cs" />
